adding blank images for sidebar and sidebar icon since setting them to NULL causes screen to not load
after scene optimizations, something changed in core libraries that is
causing scene not to load if sidebar icon and image are not set. To
prevent that I created two blank icons that can be used where we don’t
want to show button or sidebar. This is a hack and not an actual fix